refactor:useDiscountCalculator 커스텀 훅으로 만들어서 사용
