+++ src/refactoring/entities/cart/cartItem.ts
@@ -1,4 +1,4 @@
-import { Product } from "./product";
+import { Product } from "../product/product";
 
 export interface CartItem {
   product: Product;
