+++ src/refactoring/entities/discount/discount-input-form.tsx
@@ -1,5 +1,5 @@
+import { Button } from "../../ui/button";
 import { Input } from "../../ui/input";
-import { Button } from "../ui/Button";
 
 type DiscountInput = {
   quantity: number;
