+++ src/refactoring/models/cart.ts
@@ -1,5 +1,5 @@
 /**테스트 코드 통과 위해 export해옴 */
 export { calculateItemTotal } from "../calculations/cart/calc-item";
 export { getMaxApplicableDiscount } from "../calculations/discount/calc-discount-rate";
+export { calculateCartTotal } from "../calculations/discount/calc-total-discount";
 export { updateQuantity as updateCartItemQuantity } from "../logic/cart/cart-logic";
-export { calculateCartTotal } from "./../calculations/cart/calc-total-discount";
